Moved advection interp. to AdvectionTerm
Previously, some (but not all) routines for the advection interpolation
coefficients were kept in the AdvectionDiffusionTerm class (presumably
because the interpolation coefficients can also depend on the diffusion
coefficients, which only are accessible from that class). They have now
been moved to the AdvectionTerm class instead so that all routines are
contained in it. The overall interface in AdvectionDiffusionTerm doesn't
change however, since it is derived from the AdvectionTerm class.

(The main reason for moving the code is to make the
SetAdvectionInterpolationBoundaryConditions() method accessible directly
in the 'GeneralAdvectionTerm' class in the C++ tests).
