+++ include/FVM/Equation/AdvectionDiffusionTerm.hpp
@@ -15,16 +15,6 @@ namespace DREAM::FVM {
         std::vector<AdvectionTerm*> advectionterms;
         std::vector<DiffusionTerm*> diffusionterms;
 
-        enum AdvectionInterpolationCoefficient::adv_interpolation advectionInterpolationMethod_r  = AdvectionInterpolationCoefficient::AD_INTERP_CENTRED;
-        enum AdvectionInterpolationCoefficient::adv_interpolation advectionInterpolationMethod_p1 = AdvectionInterpolationCoefficient::AD_INTERP_CENTRED;
-        enum AdvectionInterpolationCoefficient::adv_interpolation advectionInterpolationMethod_p2 = AdvectionInterpolationCoefficient::AD_INTERP_CENTRED;
-        
-        real_t fluxLimiterDampingFactor = 1.0;
-
-        // The following set of variables are used for dynamic damping of flux limiters
-        bool withDynamicFluxLimiterDamping = true;
-        real_t dampingWithIteration = 1.0, t_prev = -1.0, dt_prev = -1.0;
-        len_t iteration=0;
     public:
         AdvectionDiffusionTerm(Grid *g)
             : AdvectionTerm(g, true), DiffusionTerm(g, true) {}
@@ -37,8 +27,6 @@ namespace DREAM::FVM {
 
         virtual void Rebuild(const real_t, const real_t, UnknownQuantityHandler*) override;
         virtual void ResetCoefficients() override;
-        void RebuildInterpolationCoefficients(UnknownQuantityHandler*);
-
 
         virtual void SetJacobianBlock(
             const len_t uqtyId, const len_t derivId, Matrix *jac, const real_t *x
@@ -75,35 +63,6 @@ namespace DREAM::FVM {
         virtual void SaveCoefficientsSFile(const std::string&) override;
         virtual void SaveCoefficientsSFile(SFile*) override;
         
-        // set the interpolation
-        void SetAdvectionInterpolationMethod(
-            AdvectionInterpolationCoefficient::adv_interpolation intp, 
-            FVM::fluxGridType fgType, len_t id, real_t damping_factor 
-        ){
-            this->fluxLimiterDampingFactor = damping_factor;
-            if(fgType == FLUXGRIDTYPE_RADIAL){
-                this->advectionInterpolationMethod_r = intp; 
-                this->deltar->SetUnknownId(id);
-            } else if(fgType == FLUXGRIDTYPE_P1){
-                this->advectionInterpolationMethod_p1 = intp;
-                this->delta1->SetUnknownId(id);
-            } else if(fgType == FLUXGRIDTYPE_P2){
-                this->advectionInterpolationMethod_p2 = intp;
-                this->delta2->SetUnknownId(id);
-            } 
-        }
-        void SetAdvectionBoundaryConditions(
-            fluxGridType fgType, AdvectionInterpolationCoefficient::adv_bc bc_lower, 
-            AdvectionInterpolationCoefficient::adv_bc bc_upper
-        ){
-            if(fgType == FLUXGRIDTYPE_RADIAL)
-                this->deltar->SetBoundaryConditions(bc_lower,bc_upper);
-            else if(fgType == FLUXGRIDTYPE_P1)
-                this->delta1->SetBoundaryConditions(bc_lower,bc_upper);
-            else if(fgType == FLUXGRIDTYPE_P2)
-                this->delta2->SetBoundaryConditions(bc_lower,bc_upper);
-        }
-
     };
 }
